+++ src/Events/TaskReceived.php
@@ -4,14 +4,14 @@ namespace Itinysun\Laraman\Events;
 
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Events\Dispatchable;
-use Illuminate\Http\Request;
-class MessageReceived
+
+class TaskReceived
 {
     use Dispatchable;
     public function __construct(
         public Application $app,
         public Application $sandbox,
-        public Request $request
+        public $data
     ) {
     }
 }