+++ src/Listeners/CleanBaseState.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Itinysun\Laraman\Listeners;
+
+class CleanBaseState
+{
+    public function handle($event): void
+    {
+        $event->app->cleanBaseState();
+    }
+}