+++ changelog.md
@@ -2,7 +2,13 @@
 
 All notable changes to `Laraman` will be documented in this file.
 
-## Version 0.0.3
+## version 0.0.5
+- move methods to traits
+- add event-listener mode
+- add dcat-admin support
+- fix bug when response give an unknown statue-code 
+
+## Version 0.0.4
 
 ### feat
 - add clean mode to support unknown app